fixed bugs,cleaned up code
fixed friend request bugs and made code cleaner

+++ src/main/java/models/FriendRequestModel.java
@@ -1,5 +1,7 @@
 package models;
 import domain.FriendRequest;
+import enums.ChangeEvent;
+import exceptions.ValidationException;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import repo.DbFriendRequestRepo;
@@ -12,6 +14,7 @@ import java.util.Optional;
 public class FriendRequestModel extends Observable {
 
     private final DbFriendRequestRepo friendRequestRepo;
+    private FriendRequest lastFriendRequestSaved;
 
     public FriendRequestModel(DbFriendRequestRepo friendRequestRepo) {
         this.friendRequestRepo = friendRequestRepo;
@@ -46,15 +49,25 @@ public class FriendRequestModel extends Observable {
 
     public void save(FriendRequest friendRequest){
         friendRequestRepo.save(friendRequest);
-        notifyObservers();
+        notifyObservers(ChangeEvent.SENT_FRIEND_REQUEST);
     }
     public void delete(Long id1, Long id2){
+        if(id1<0 || id2<0){
+            throw new ValidationException("Id's cannot be negative");
+        }
         friendRequestRepo.delete(new Tuple<>(id1,id2));
-        notifyObservers();
+        notifyObservers(ChangeEvent.FRIEND_REQUEST_DATA);
     }
     public void update(FriendRequest friendRequest){
         friendRequestRepo.update(friendRequest);
-        notifyObservers();
+        notifyObservers(ChangeEvent.USER_DATA);
     }
 
+    public FriendRequest getLastFriendRequest() {
+        return lastFriendRequestSaved;
+    }
+
+    public void setLastFriendRequestSaved(FriendRequest lastFriendRequestSaved) {
+        this.lastFriendRequestSaved = lastFriendRequestSaved;
+    }
 }
